Change frontend to async await syntax
* Use tonic instead of tower-grpc
* Use warp instead of tower-web (not sure, but
  warp seems to have better support?)
