+++ frontendservice/Cargo.toml
@@ -9,23 +9,22 @@ name = "frontend-server"
 path = "src/main.rs"
 
 [dependencies]
+async-stream = "0.2"
 bytes = "0.4"
-futures = "0.1.18"
-http = "0.1"
+futures = { version = "0.3", default-features = false, features = ["alloc"]}
+http = "0.2"
 log = "0.4"
 pretty_env_logger = "0.3.1"
 prost = "0.5"
-tokio = "0.1.10"
+tokio = { version = "0.2", features = ["rt-threaded", "time", "stream", "fs", "macros", "uds"] }
+tonic = { git = "https://github.com/hyperium/tonic", features = ["tls"] }
 tower = { git = "https://github.com/tower-rs/tower" }
-tower-add-origin = { git = "https://github.com/tower-rs/tower-http" }
-tower-buffer = { git = "https://github.com/tower-rs/tower" }
-tower-grpc = { git = "https://github.com/tower-rs/tower-grpc" }
-tower-h2 = { git = "https://github.com/tower-rs/tower-h2" }
 tower-service = { git = "https://github.com/tower-rs/tower" }
-tower-web = { git = "https://github.com/carllerche/tower-web", default-features = false }
+# TODO: disable default-features
+warp = { git = "https://github.com/seanmonstar/warp" }
 
 [build-dependencies]
-tower-grpc-build = { git = "https://github.com/tower-rs/tower-grpc" }
+tonic-build = { git = "https://github.com/hyperium/tonic" }
 
 [patch.crates-io]
 tower-service = { git = "https://github.com/tower-rs/tower" }
