+++ frontendservice/Dockerfile
@@ -6,7 +6,11 @@ ENV FORTUNE_SERVICE_HOSTNAME fortuneservice
 
 COPY . /code
 WORKDIR /code
+
+# required for tonic-build
+RUN rustup component add rustfmt
 RUN cargo build
+
 EXPOSE 8080
 
 FROM debian:buster-slim
